+++ cleanArray.f90
@@ -62,11 +62,11 @@ Module cleanArray
             integer maxpoints, II, JJ, KK, VV, ic, jc, jcc, icc, fill, kc, ff, total
             integer pointsHpos, pointsHneg, pointsVpos, pointsVneg, points, INCREASED, orgVV, orgKK, TRIES
 
-            double precision, dimension(JJ,II) :: dataout,datain
-            double precision, dimension(JJ,II) :: mask
-            double precision, dimension(4,3,maxpoints)  :: foundWeight, foundData
-            double precision counter, hori, vert, sumweights, distance, sodadepth, weight
-            double precision, dimension(JJ,II) :: bathymetry
+            real(4), dimension(JJ,II) :: dataout,datain
+            real(4), dimension(JJ,II) :: mask
+            real(4), dimension(4,3,maxpoints)  :: foundWeight, foundData
+            real(4) counter, hori, vert, sumweights, distance, sodadepth, weight
+            real(4), dimension(JJ,II) :: bathymetry
             
 !f2py intent(in,out,overwrite) dataout
 !f2py intent(in,overwrite) datain, mask, bathymetry
